+++ deployments/ansible_collections/signalfx/splunk_otel_collector/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### 💡 Enhancements 💡
+
+- Recursively set the ownership of collector config (`/etc/otel/collector`) and state (`/var/lib/otelcol`) directories to the configured service user and group on Linux.
+
 ## ansible-v1.1.0
 
 ### 🛑 Breaking changes 🛑